app/pyproject.toml
[project]
name = "wallet-tracker-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "flask-sqlalchemy>=3.1.1",
    "pyjwt>=2.12.0",
    "mysqlclient>=2.2.8",
    "cryptography>=46.0.6",
    "requests>=2.33.0",
    "flask-migrate>=4.0.7",
    "waitress>=3.0.2; sys_platform == 'win32'",
    "uwsgi>=2.0.28; sys_platform != 'win32'",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-flask>=1.3.0",
    "pytest-mock>=3.15.1",
    "pytest-env>=1.5.0",
    "pip-audit>=2.9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
env_files = [".env"]
